backend/requirements.txt
pillow==11.1.0
unstructured==0.15.13
fastapi==0.115.0
chardet==5.2.0
pypdf==5.2.0
openai==1.64.0
anthropic==0.47.1
google-genai==1.9.0
uvicorn==0.34.0
lxml[html_clean]
python-magic; platform_system != "Windows"
python-magic-bin; platform_system == "Windows"
aiohttp==3.11.11
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.20
pytest-playwright==0.7.0
playwright-stealth==1.0.6
nbconvert==7.16.6
markdownify==0.14.1
# Add Database and Encryption related libraries
SQLAlchemy>=1.4,<2.0 # Pin major version for stability
psycopg2-binary # For PostgreSQL connection
cryptography # For encrypting API keys
alembic # For database migrations
